Changed integer PKs to be bigints instead

+++ crates/database/src/feed_entries.rs
@@ -1,5 +1,5 @@
 #[derive(Debug)]
 pub struct InsertData {
-    pub feed_id: i32,
-    pub entry_id: i32,
+    pub feed_id: i64,
+    pub entry_id: i64,
 }
