+++ crates/database/src/profile_feed_entries.rs
@@ -1,5 +1,5 @@
 #[derive(Debug)]
 pub struct InsertData<'a> {
     pub profile_feed_id: &'a str,
-    pub feed_entry_id: i32,
+    pub feed_entry_id: i64,
 }
